Consolida nodos en un único nodo VictoriaOS con Trigger
Cambios principales:
- Crea nodo VictoriaOS que consolida todas las operaciones (Tareas, Webhooks, Usuario)
- Crea nodo VictoriaOsTrigger para escuchar eventos y comenzar automatizaciones
- Elimina nodos separados (Tareas, Usuario, Webhooks)
- Incrementa versión a 1.1.0

El nodo principal ahora incluye:
- Recurso Tareas: listar, obtener, crear, actualizar, eliminar
- Recurso Webhooks: listar, obtener, crear, actualizar, eliminar, estadísticas
- Recurso Usuario: obtener información del usuario

El Trigger soporta eventos:
- task.created, task.updated, task.completed, task.deleted
